merge

+++ index.js
@@ -38,15 +38,17 @@ client.on("messageCreate", async (message) => {
         }
     }
 
-    if (message.content.toLowerCase() == "!flipcoin") {
-        const coin = Math.random() < 0.5 ? "heads" : "tails";
-        message.reply(coin);
-    }
-
 });
 
 client.on('interactionCreate', async (interaction) => {
     if (!interaction.isChatInputCommand()) return;
+
+    if (interaction.commandName == "coin_flip") {
+        if (Math.random() < 0.5) {
+            interaction.reply("head")
+        }
+        else interaction.reply("tail")
+    }
 
     if (interaction.commandName == "vid") {
         let url = interaction.options.get('url').value;
@@ -112,22 +114,34 @@ client.on('interactionCreate', async (interaction) => {
         // insta reels / fb post
         if (url.includes("instagram.com/reel") || url.includes("www.facebook.com")) {
             interaction.reply("downloading video....")
+            let video
 
             try {
                 const download_url = await get_insta_download_url(url)
+                if (!download_url) {
+                    interaction.editReply("unable to access url")
+                    return
+                }
                 const vid_file = './output/insta_reel.mp4'
-                const video = await download_file_from_url(download_url, vid_file)
-                console.log("download url: " + video)
-                if (video) {
-                    const file = await get_vid(vid_file);
-                    await interaction.editReply({
-                        content: `here's ur vid bud <@${interaction.user.id}>`,
-                        files: [{
-                            attachment: file,
-                            contentType: "video/mp4",
-                            name: "fked_mc_download.mp4",
-                    }] });
-                    delete_file(vid_file)
+
+                try {
+                    video = await download_file_from_url(download_url, vid_file)
+                    if (video) {
+                        const file = await get_vid(vid_file);
+                        await interaction.editReply({
+                            content: `here's ur vid bud <@${interaction.user.id}>`,
+                            files: [{
+                                attachment: file,
+                                contentType: "video/mp4",
+                                name: "fked_mc_download.mp4",
+                        }] });
+                        delete_file(vid_file)
+                    }
+                }
+                catch (error) {
+                    console.error("error: ", error)
+                    interaction.editReply("an issue occured while downloading video")
+                    return
                 }
 
             } catch (error) {
@@ -135,7 +149,7 @@ client.on('interactionCreate', async (interaction) => {
                 if (error.rawError.message == "Request entity too large") {
                     interaction.editReply("video exceed file size limit")
                 }
-                else interaction.editReply("an issue occured while downloading video")
+                else interaction.editReply("error : (")
             }
 
             return
@@ -146,6 +160,10 @@ client.on('interactionCreate', async (interaction) => {
 
             try {
                 const download_url = await get_tiktok_download_url(url)
+                if (!download_url) {
+                    interaction.editReply("unable to access url")
+                    return
+                }
                 const vid_file = './output/tik_tok.mp4'
                 const video = await download_file_from_url(download_url, vid_file)
                 console.log("download url: " + video)
@@ -181,6 +199,10 @@ client.on('interactionCreate', async (interaction) => {
 
             try {
                 const download_url = await get_twitter_download_url(url)
+                if (!download_url) {
+                    interaction.editReply("unable to access url")
+                    return
+                }
                 const vid_file = './output/tweet.mp4'
                 const video = await download_file_from_url(download_url, vid_file)
                 console.log("download url: " + video)
